refactoring e2e testsuite code to group together logically (#3388)
